# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
...
  /api/v1/salary_slips:
    get:
      tags:
        - SalarySlips
      summary: List SalarySlips
      description: Get a paginated list of SalarySlips with optional filtering and sorting
      operationId: listSalarySlips
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SalarySlipList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - SalarySlips
      summary: Create SalarySlip
      description: Create a new SalarySlip
      operationId: createSalarySlip
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSalarySlipRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SalarySlip'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/salary_components:
    get:
      tags:
        - SalaryComponents
      summary: List SalaryComponents
      description: Get a paginated list of SalaryComponents with optional filtering and sorting
      operationId: listSalaryComponents
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SalaryComponentList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - SalaryComponents
      summary: Create SalaryComponent
      description: Create a new SalaryComponent
      operationId: createSalaryComponent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSalaryComponentRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SalaryComponent'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
...
    SalarySlip:
      type: object
      required:
        - id
        - payroll_entry_id
        - company_id
        - employee_id
        - working_days
        - unpaid_days
        - gross_pay
        - total_deductions
        - net_pay
        - metadata
      properties:
        id:
          type: string
          format: uuid
        payroll_entry_id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        employee_id:
          type: string
          format: uuid
        structure_id:
          nullable: true
          type: string
          format: uuid
        working_days:
          type: string
        unpaid_days:
          type: string
        gross_pay:
          type: string
        total_deductions:
          type: string
        net_pay:
          type: string
        overtime_hours:
          nullable: true
          type: string
        tax_method:
          nullable: true
          type: string
          maxLength: 20
        metadata:
          type: object
...
    CreateSalarySlipRequest:
      type: object
      required:
        - payroll_entry_id
        - company_id
        - employee_id
        - working_days
        - unpaid_days
        - gross_pay
        - total_deductions
        - net_pay
        - metadata
      properties:
        payroll_entry_id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        employee_id:
          type: string
          format: uuid
        structure_id:
          nullable: true
          type: string
          format: uuid
        working_days:
          type: string
        unpaid_days:
          type: string
        gross_pay:
          type: string
        total_deductions:
          type: string
        net_pay:
          type: string
        overtime_hours:
          nullable: true
          type: string
        tax_method:
          nullable: true
          type: string
          maxLength: 20
        metadata:
          type: object
...
    SalarySlipList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/SalarySlip'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    SalaryComponent:
      type: object
      required:
        - id
        - structure_id
        - company_id
        - name
        - component_type
        - amount
        - gl_account_id
        - metadata
      properties:
        id:
          type: string
          format: uuid
        structure_id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        name:
          type: string
          maxLength: 140
        component_type:
          type: string
        amount:
          type: string
        gl_account_id:
          type: string
          format: uuid
        metadata:
          type: object
...
    CreateSalaryComponentRequest:
      type: object
      required:
        - structure_id
        - company_id
        - name
        - component_type
        - amount
        - gl_account_id
        - metadata
      properties:
        structure_id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        name:
          type: string
          maxLength: 140
        component_type:
          type: string
        amount:
          type: string
        gl_account_id:
          type: string
          format: uuid
        metadata:
          type: object
...
    SalaryComponentList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/SalaryComponent'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []